# docs/swagger.yaml
basePath: /
definitions:
  controllers.loginPayload:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  controllers.signupPayload:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    type: object
  models.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updatedAte:
        type: string
    type: object
  structs.ResponseBody:
    properties:
      code:
        type: string
      data: {}
      msg:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: nocvi111@gmail.com
  description: This is a sample forum server.
  license:
    name: MIT
    url: https://github.com/shlason/go-forum/blob/main/LICENSE
  title: Go-forum API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: Login by email or name and password
      parameters:
      - description: email or name and password
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/controllers.loginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ResponseBody'
      summary: Login
      tags:
      - auth
  /signup:
    post:
      consumes:
      - application/json
      description: Create account by email, name, password
      parameters:
      - description: email, name, password
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/controllers.signupPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/structs.ResponseBody'
            - properties:
                data:
                  $ref: '#/definitions/models.User'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/structs.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/structs.ResponseBody'
      summary: Create account
      tags:
      - auth
swagger: "2.0"